backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
chromadb
sentence-transformers
scikit-learn
langgraph
groq
requests
httpx
redis
celery
prometheus-client
transformers
ultralytics
